Improve `ruff_python_semantic::all::extract_all_names()` (#11335)
