+++ crates/ruff_linter/resources/test/fixtures/pyflakes/F822_1b.py
@@ -1,0 +1,4 @@
+import builtins
+a = 1
+
+__all__ = builtins.list(["a", "b"])
