+++ crates/ruff_python_semantic/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod all;
 pub mod analyze;
 mod binding;
 mod branches;
