+++ crates/ruff_python_semantic/src/model.rs
@@ -1,3 +1,5 @@
+pub mod all;
+
 use std::path::Path;
 
 use bitflags::bitflags;
